refactor: UI 関連のスクリプトを分離
