+++ Assets/MyGame/Scripts/StageControlScript.cs
@@ -14,12 +14,6 @@ public class StageControlScript : MonoBehaviour
     public GameObject SettingUI;
     public GameObject LifePiUI;
 
-    // ライフのアイコンを格納
-    public GameObject[] lifeIcons;
-
-    // Pi取得数
-    public Text piCount;
-
     // プレイヤーコントローラーの取得
     PlayerControlScript playerControlScript;
 
@@ -34,10 +28,6 @@ public class StageControlScript : MonoBehaviour
 
     void LateUpdate()
     {
-        // ライフとPiのUIを更新
-        UpdateLife();
-        UpdatePiCount();
-
         switch (state)
         {
             case State.Ready:
@@ -139,30 +129,6 @@ public class StageControlScript : MonoBehaviour
     public void OnToPauseButtonClicked()
     {
         Pause();
-    }
-
-    // ライフの表示数を管理
-    private void UpdateLife()
-    {
-        int life = playerControlScript.getLife();
-        for (int i = 0; i < lifeIcons.Length; i++)
-        {
-            if (i < life)
-            {
-                lifeIcons[i].SetActive(true);
-            }
-            else
-            {
-                lifeIcons[i].SetActive(false);
-            }
-        }
-    }
-
-    // Piの取得数を管理
-    private void UpdatePiCount()
-    {
-        int pi = playerControlScript.getPi();
-        piCount.text = pi + " / 8";
     }
 
     public static string getSceneName()
