+++ Assets/MyGame/Scripts/UIControlScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 55d29956dee6873dea080e8d0644a2d9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
